Changed user to root
Creator hardcoded his Linux username into the service so for people with accounts not named "tim" the service will fail to start.  As I am using this on RasPBX, which uses the root account, that is what it's being changed to.  Looking into being able to run this as something other than root for security reasons,